refactor: auth, member 패키지 수정
