+++ backend/main-service/src/main/java/kkakka/mainservice/common/config/WebMvcConfig.java
@@ -1,10 +1,12 @@
-package kkakka.mainservice.auth.infrastructure.config;
+package kkakka.mainservice.common.config;
 
+import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@AllArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
 
     @Override
