+++ backend/main-service/src/main/java/kkakka/mainservice/member/member/domain/Grade.java
@@ -1,4 +1,4 @@
-package kkakka.mainservice.member.domain;
+package kkakka.mainservice.member.member.domain;
 
 /*
  * TODO: Grade enum -> entity 로 변경할 것
