+++ backend/main-service/src/main/java/kkakka/mainservice/member/member/domain/Member.java
@@ -1,4 +1,4 @@
-package kkakka.mainservice.member.domain;
+package kkakka.mainservice.member.member.domain;
 
 import java.util.Objects;
 import javax.persistence.Embedded;
