+++ backend/main-service/src/main/java/kkakka/mainservice/member/member/ui/MemberController.java
@@ -1,11 +1,11 @@
-package kkakka.mainservice.member.ui;
+package kkakka.mainservice.member.member.ui;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/member")
+@RequestMapping("/members")
 public class MemberController {
 
     @GetMapping("/health_check")
