+++ backend/main-service/src/main/java/kkakka/mainservice/member/member/ui/dto/MemberResponse.java
@@ -1,0 +1,4 @@
+package kkakka.mainservice.member.member.ui.dto;
+
+public class MemberResponse {
+}
