--- backend/main-service/src/main/java/kkakka/mainservice/member/ui/dto/MemberResponse.java
@@ -1,4 +1,0 @@
-package kkakka.mainservice.member.ui.dto;
-
-public class MemberResponse {
-}
