--- backend/main-service/src/test/java/kkakka/mainservice/MainServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package kkakka.mainservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MainServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
